style: 파라미터 개행 관련 코드 컨벤션 적용
- 파라미터의 개수가 2개 이상인 경우 개행
- 클로저 내부 파라미터의 경우 미적용
