+++ OpenMarket/Mock/MockURLSession.swift
@@ -16,9 +16,12 @@ class MockURLSession: URLSessionProtocol {
         self.response = response
     }
     
-    func dataTask(with url: URL, completion: @escaping (Data?, URLResponse?, Error?) -> Void) -> URLSessionDataTaskProtocol {
+    func dataTask(with url: URL,
+                  completion: @escaping (Data?, URLResponse?, Error?) -> Void) -> URLSessionDataTaskProtocol {
         return MockURLSessionDataTask {
-            completion(self.response.data, self.response.urlResponse, self.response.error)
+            completion(self.response.data,
+                       self.response.urlResponse,
+                       self.response.error)
         }
     }
 }
